Add files via upload
This project is a work in progress. The concept is a loan calculator that will allow the user to input their student or other loan information and use a slider to see how various repayment amounts will impact total interest and total amount paid. The goal is to help users understand how small increases in additional payment amounts can lead to larger long-term savings in total paid.